refactor: use generic FakeStreamRegistry in Publisher.test.ts
